Equalize features between interop and service, code refactor
